fix: remove mongoDB credentials from repository
They should live in the .env file and this file should not be committed to repo.
In a production app, we should clean the entire git history following this kind of tutorial
https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/removing-sensitive-data-from-a-repository
Also we could think that this DB is compromised and it would be better deleting it and creating another one.