fix: unique index on project name
There is a unique index on the name for the Project model. This is bad because if 2 users, from different organisations, create a project with the same name, the second one will have an error "project already taken" but will not see it in his projects list (because project list is scoped to the current user’s organisation)
We should change it to a compound unique index based on name+organisation.
In real life we should also resync indexes to remove the "project name index" but because the DB is shared with other candidates I won't do it here.